one config file, but its a mess
It is now using different dbs for data and request caching
