--- src/main/java/uk/nhs/hee/tis/trainee/sync/config/RedisLettuceConfig.java
@@ -1,65 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright 2021 Crown Copyright (Health Education England)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
- * associated documentation files (the "Software"), to deal in the Software without restriction,
- * including without limitation the rights to use, copy, modify, merge, publish, distribute,
- * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or
- * substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
- * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
- * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package uk.nhs.hee.tis.trainee.sync.config;
-
-import io.lettuce.core.ClientOptions;
-import io.lettuce.core.RedisClient;
-import io.lettuce.core.TimeoutOptions;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-
-@Configuration
-public class RedisLettuceConfig {
-
-  @Value("${spring.redis.host}")
-  String host;
-
-  @Value("${spring.redis.port}")
-  Integer port;
-
-  @Value("${spring.redis.password}")
-  String password;
-
-  @Value("${spring.redis.url}")
-  String redisUrl;
-
-  @Bean
-  public LettuceConnectionFactory redisConnectionFactory() {
-    RedisStandaloneConfiguration redisConf = new RedisStandaloneConfiguration(host, port);
-    redisConf.setPassword(password);
-    //set timeout?
-    return new LettuceConnectionFactory(redisConf);
-  }
-
-  @Bean
-  public RedisClient getRedisClient() {
-//    TimeoutOptions timeoutOptions = TimeoutOptions.create().
-//    ClientOptions clientOptions = ClientOptions.create().getTimeoutOptions();
-
-    RedisClient redisClient = RedisClient.create(redisUrl);
-//    redisClient.setOptions();
-    return redisClient;
-  }
-}
